backend/requirements.txt
# FastAPI Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1

# Pydantic & Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# ML & Data Science
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# LLM Integration
google-generativeai==0.3.2

# Data Visualization & Export
matplotlib==3.8.2
seaborn==0.13.2

# Utilities
python-dateutil==2.8.2
faker==22.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Development Tools
black==24.1.1
flake8==7.0.0
